- changed storage on zfs to rely on zfs-snapshots for retension
On non-ZFS storage we utilize hard-links and rsync's link-dest option
for retaining previous backups and keeping current backup as a full.
The hardlinking give us file-level storage efficiency within a backup
module.

On ZFS we don't need to do this, as using snapshots preserves previous
backups, and as long as dedup is enabled on the ZFS backup datastore
then we will have block-level storage efficiency, and it will be across
all modules.